feat: add methods `uploadResourceAsBlob` and `downloadResourceAsBlob`
