+++ package.json
@@ -37,7 +37,9 @@
     "eslint": "eslint . --ext ts",
     "prettier": "prettier \"**/*.{css,html,ts,js,java}\"",
     "swiftlint": "node-swiftlint",
-    "docgen": "docgen --api CloudinaryPlugin --output-readme README.md --output-json dist/docs.json",
+    "docgen": "npm run docgen:plugin && npm run docgen:utils",
+    "docgen:plugin": "docgen --api CloudinaryPlugin --output-readme README.md --output-json dist/docs.json",
+    "docgen:utils": "docgen --api ICloudinaryUtils --output-readme docs/utils/README.md --output-json dist/utils-docs.json",
     "build": "npm run clean && npm run docgen && tsc && rollup -c rollup.config.js",
     "clean": "rimraf ./dist",
     "watch": "tsc --watch",
